Allow to apply opacity to all background colors
Alacritty currently only applies the value of `background_opacity` to
the default background and not to all (e.g. #807, addressed in #847).
However this is not necessarily always desired (e.g. #1082, #1180).

This patch adds a configuration option and logic (by partly re-instating
code removed in #847) to toggle the behaviour between either applying
only to the default background color, or all background colors.  For
example iterm2 also has this option and the naming of the toggle is
borrowed from there.